Split up main version_check.py script into more files
