+++ version_check/config/__init__.py
@@ -1,0 +1,5 @@
+# -*- coding: utf-8 -*-
+
+VERSION = '0.1.0'
+GIT_DIR = '--git-dir=/salt/.git'
+REMOTE = 'origin'
